Fix #20
Line comment was defined as (from "//" up to line break) and, obviously, at the end of a file, there was no line break. Even if you put a line break after that comment, that line break was probably filtered away by a noise rule instead.

This fix re-defines the line comment as ("//" followed by non-line-break characters).