PRX-7 Tolerate #!/interpreter/line
Add a special hack to the lexer and parser that let them tolerate `#!<until EOL>` at the beginning of a Prexonite Script file.

The rules are not very strict at the moment (doesn't need to be the first bytes in the stream; may even show up after comments), not because it makes sense to support it but because it would be more difficult to forbid it.

I didn't want to go for a solution that hooks itself in before the lexer because that would mean introducing an entirely new mechanism/concept to the translation pipeline. With a lexer/parser based solution, I'm re-using existing infrastructure, which is going to be better in terms of maintenance long term.