PRX-43 Use `/` instead of `,`  for ModuleName
The Prexonite language uses `/` to separate module name and version. The Engine should do the same. Using a different notation doesn't serve a useful purpose.

This is a BREAKING change, though it is very unlikely that any real code is affected as module name references to be parsed are written in Prexonite Script (where the compiler handles the parsing, and not the `TryParse` method of ModuleName).