PRX-29: add warning for unqualified import
`namespace import x;` doesn't do anything. Either `x` is already in scope (in which case, this is a no-op) or `x` is not in scope (which results in an error symbol being declared).

For now, I've decided not to immediately emit an error for unresolved symbols. Not 100% sure if I'm going to keep imports this way.

The rationale is that Prexonite Script doesn't have robust IDE support and thus micro-managing imports is painful. Having a compilation fail because of an unused but unresolved import sounds annoying to me.

Also:
 - Introduce `GetNextPosition` method. See PRX-57